Initial documentation of bencode implementation.
I'm not aware of preferred documentation practices, so I'm open to
suggestions. I'm reasonably confident in the documentation's
correctness and communicability, but could use a second opinion.

The multi-value encapsulation/multi-part definition in the comments
of benc_strcmp in the header could probably look better, though I'm
trying to avoid anything too fancy.